Fix tombstone handling for tables
Changelog topics for tables can contain tombstone records with null values to
indicate that a given key is deleted from a table. KSQL implements tables by
building a KStream over the changelog topic, applying some transformations for
adding columns (one of which Kafka Streams only supports for KStreams), and
then reducing the result into a KTable. The problem is that Kafka Streams
filters out null values in the reduce and aggregation processors for
GroupedKStreams. This patch works around this problem by mapping to
Optional<GenericRow> just before grouping and aggregating into the final
KTable, and then implementing an aggregator for handling Optional.EMPTY as
a tombstone